+++ src/infrastructures/models/user.rs
@@ -4,7 +4,7 @@ use rust_api::schema::users;
 // NOTE #[derive(Queryable)] will generate all of the code needed to load a Post struct from a SQL query.
 #[derive(Queryable, PartialEq, Debug)]
 pub struct User {
-    pub id: u64,
+    pub id: i32,
     pub first_name: String,
     pub last_name: Option<String>,
     pub email: Option<String>,
@@ -13,7 +13,7 @@ pub struct User {
 #[derive(Insertable)]
 #[table_name = "users"]
 pub struct NewUser {
-    // pub id: u64,
+    // pub id: i32,
     pub first_name: String,
     pub last_name: Option<String>,
     pub email: Option<String>,
